fix(sdk): follow lint rules
