+++ packages/sdk/src/errors/BrowserWalletNotInstalledError.ts
@@ -1,0 +1,6 @@
+export class BrowserWalletNotInstalledError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "BrowserWalletNotInstalledError";
+  }
+}
