+++ packages/sdk/src/errors/BrowserWalletSigningError.ts
@@ -1,0 +1,6 @@
+export class BrowserWalletSigningError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "BrowserWalletSigningError";
+  }
+}
