+++ packages/sdk/src/errors/OrditSDKError.ts
@@ -1,0 +1,6 @@
+export class OrditSDKError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "OrditSDKError";
+  }
+}
